snapshot of project "lynx", label v2-8-8dev_16e

+++ WWW/Library/Implementation/HTTP.c
@@ -1,5 +1,5 @@
 /*
- * $LynxId: HTTP.c,v 1.130 2013/10/01 21:58:37 tom Exp $
+ * $LynxId: HTTP.c,v 1.131 2013/10/02 14:33:32 tom Exp $
  *
  * HyperText Tranfer Protocol	- Client implementation		HTTP.c
  * ==========================
@@ -717,7 +717,7 @@ static int HTLoadHTTP(const char *arg,
 	    handle->options |= SSL_OP_NO_TLSv1;
 #if OPENSSL_VERSION_NUMBER >= 0x0090806fL && !defined(OPENSSL_NO_TLSEXT)
 	} else {
-	    int ret = SSL_set_tlsext_host_name(handle, ssl_host);
+	    int ret = (int) SSL_set_tlsext_host_name(handle, ssl_host);
 
 	    CTRACE((tfp, "...called SSL_set_tlsext_host_name(%s) ->%d\n",
 		    ssl_host, ret));
